v0.6.9 - Working on core functionality.
The idea behind installation.....

Anyone can launch the UI.

There are functions for detecting if admin and restarting if not admin. These can be used to detect if the session has enough priviledges to install.

Once the app has admin rights it creates an HRAV user and adds it to the administrators group.

Then it restarts as this user using PAExec and a specified password.

The password is never stored, only the keys and salts. The password will be generated everytime it is needed from the keys and salts.

The password will be changed and the salts/keys randomized. Ideally they should rotate every 12-24hours.

By creating an admin user and running all the scripts as that user, we can accomplish a lot of background tasks that real-time protection is going to need without throwing UAC boxes all the time and interfering with users who might not even have local admin rights.